Move styles.tsx to avoid page routing conflict
