+++ styles/tsxBased/calculatorPagestyles.tsx
